# openapi.yaml
...
info:
  description: 'The operator console: a session-gated HTML surface over the deployment manifest — dashboards, configuration, branding, tenants, backups, the trust and feedback views, and the adoption journeys. It holds no domain logic: it reads the manifest and proxies the services'' own surfaces. Login requires the console admin token where one is configured.'
  license:
    identifier: Apache-2.0
    name: Apache-2.0
  title: UniDPP console
  version: 0.1.0
  x-unidpp-env-keys:
  - UNIDPP_CONSOLE_BIND
  - UNIDPP_CONSOLE_MANIFEST
  - UNIDPP_CONSOLE_ADMIN_TOKEN
openapi: 3.1.0
paths:
...
  /passports:
    get:
      operationId: passports_page
      responses:
        '200':
          content:
            text/html:
              schema:
                type: string
          description: The page, rendered against the deployment manifest
        '303':
          description: The session gate redirects an unauthenticated browser to `/login`
      summary: The verification form and its most recent verdicts.
      tags:
      - console
    post:
      operationId: submit
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: string
        description: 'The verification form: the pack reference and its anchors'
        required: true
      responses:
        '200':
          content:
            text/html:
              schema:
                type: string
          description: The form is re-rendered with its error stated
        '303':
          description: 'Verified: a redirect back to the form with the verdict stated'
      summary: 'POST /passports with a pack: verify through the pipeline.'
      tags:
      - console
...